requirements.txt
# Requirements inferred from imports in src/
pandas
numpy
scipy
yfinance
pyyaml

# Development dependencies
pytest
ib_insync
